Initial Commit for Wombats re-write
The goal here is to bring Wombats up to speed with current best
pratices in the Clojure community and leverage technologies that
play well in the Clojure ecosystem.

- Pedestal
- Datomic
- core.spec
- core.async
- Component

In addition to these changes we want to take a step back and look
at the work already done to see where we can make improvements
to modularize it a bit more.